requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==3.2.10
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-spectacular==0.27.1
exceptiongroup==1.2.0
Faker==12.0.1
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.5
MarkupSafe==2.1.5
mixer==7.2.2
oauthlib==3.2.2
packaging==24.0
pluggy==1.4.0
prompt-toolkit==3.0.43
pycparser==2.21
PyJWT==2.8.0
pyotp==2.9.0
pytest==8.1.1
pytest-django==4.8.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
referencing==0.34.0
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.5.3
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
